google map key updates (#1165)
* Update GoogleMap ApiKey to use placeholder value

Changed the `ApiKey` in `appsettings.json` from a specific key to a placeholder (`"<<ApiKey>>"`). This update enhances security by preventing the exposure of sensitive information in the source code.

* Update Google Maps API key link in documentation

The link for obtaining a Google Maps API key has been updated to direct users to the correct page specifically for getting an API key, improving the relevance for prerequisites.

NOTE: This commit message is auto-generated using GitHub Copilot.