Upgrade .NET SDK to 10.0.x in CI workflows
Updated the .NET Core SDK version from 9.0.x to 10.0.x in
`static-web-app.yml` and `web-app.yml` files. This includes
changes to the `Setup .NET Core SDK` and `Build And Deploy`
steps in `static-web-app.yml`, as well as the `Set up .NET`
step in `web-app.yml`. Removed the setup for .NET 9.0.x in
`web-app.yml` to fully transition to .NET 10.0.x.